# monkey patch with scalaz.IList