--- a/scala/bench/src/main/scala/svc/EqualBench.scala
+++ b/scala/bench/src/main/scala/svc/EqualBench.scala
@@ -18,27 +18,31 @@ class EqualBench {
   var list1: List[Int] = _
   var listK1: List[Kitties.Foo] = _
   var listK2: List[Kitties.Foo] = _
-  var listZ1: List[Zed.Foo] = _
-  var listZ2: List[Zed.Foo] = _
+  var listz0: scalaz.IList[Int] = _
+  var listz1: scalaz.IList[Int] = _
+  var listZ1: scalaz.IList[Zed.Foo] = _
+  var listZ2: scalaz.IList[Zed.Foo] = _
   var listV1: List[Bar] = _
   var listV2: List[Bar] = _
   var arr0:  Array[Int] = _
   var arrK:  Array[Kitties.Foo] = _
-  var arrZ:  Array[Zed.Foo] = _
+  var arrZ:  scalaz.ImmutableArray[Zed.Foo] = _
   var arrV:  Array[Bar] = _
 
   @Setup
   def setup(): Unit = {
     arrK  = Array.range(1, 1000).map(_ => Kitties.Foo(1000, "hello therE", true))
-    arrZ  = Array.range(1, 1000).map(_ => Zed.Foo(1000, "hello therE", true))
+    arrZ  = scalaz.ImmutableArray.fromArray(Array.range(1, 1000).map(_ => Zed.Foo(1000, "hello therE", true)))
     arrV  = Array.range(1, 1000).map(_ => Bar(1000, "hello therE", true))
     arr0  = Array.range(1, 1000)
     list0 = List.range(1, 1000)
     list1 = List.range(1, 1000)
     listK1 = arrK.toList
     listK2 = arrK.toList
-    listZ1 = arrZ.toList
-    listZ2 = arrZ.toList
+    listz0 = scalaz.IList.fromList(List.range(1, 1000))
+    listz1 = scalaz.IList.fromList(List.range(1, 1000))
+    listZ1 = scalaz.IList.fromList(arrZ.toList)
+    listZ2 = scalaz.IList.fromList(arrZ.toList)
     listV1 = arrV.toList
     listV2 = arrV.toList
   }
@@ -77,9 +81,9 @@ class EqualBench {
   @Benchmark
   def equalDiffIntVanilla: Boolean = list0 == list1
   @Benchmark
-  def equalWhileIntVanilla: Boolean = whileInt(arr0)
+  def equalWhileIntArrayVanilla: Boolean = whileInt(arr0)
   @Benchmark
-  def equalWhileClassVanilla: Boolean = whileClass(arrV)
+  def equalWhileClassArrayVanilla: Boolean = whileClass(arrV)
 
   @Benchmark
   def equalSameIntCats: Boolean = Kitties.equalAll(list0, list0)
@@ -90,21 +94,21 @@ class EqualBench {
   @Benchmark
   def equalDiffClassCats: Boolean = Kitties.equalAll(listK1, listK2)
   @Benchmark
-  def equalWhileIntCats: Boolean = Kitties.equalWhileInt(arr0)
+  def equalWhileIntArrayCats: Boolean = Kitties.equalWhileInt(arr0)
   @Benchmark
-  def equalWhileClassCats: Boolean = Kitties.equalWhileClass(arrK)
+  def equalWhileClassArrayCats: Boolean = Kitties.equalWhileClass(arrK)
 
   @Benchmark
-  def equalSameIntScalaz: Boolean = Zed.equalAll(list0, list0)
+  def equalSameIntScalaz: Boolean = Zed.equalAll(listz0, listz0)
   @Benchmark
   def equalSameClassScalaz: Boolean = Zed.equalAll(listZ1, listZ1)
   @Benchmark
-  def equalDiffIntScalaz: Boolean = Zed.equalAll(list0, list1)
+  def equalDiffIntScalaz: Boolean = Zed.equalAll(listz0, listz1)
   @Benchmark
   def equalDiffClassScalaz: Boolean = Zed.equalAll(listZ1, listZ2)
   @Benchmark
-  def equalWhileIntScalaz: Boolean = Zed.equalWhileInt(arr0)
+  def equalWhileIntArrayScalaz: Boolean = Zed.equalWhileInt(arr0)
   @Benchmark
-  def equalWhileClassScalaz: Boolean = Zed.equalWhileClass(arrZ)
+  def equalWhileClassArrayScalaz: Boolean = Zed.equalWhileClass(arrZ)
 
 }