--- a/scala/src/main/scala/svc/Zed.scala
+++ b/scala/src/main/scala/svc/Zed.scala
@@ -87,7 +87,7 @@ object Zed {
     */
   def equalOpt: Boolean = some(5) === some(6)  // false
 
-  def equalAll[A: Equal](l0: List[A], l1: List[A]): Boolean = l0 === l1
+  def equalAll[A: Equal](l0: IList[A], l1: IList[A]): Boolean = l0 === l1
 
   def equalWhileInt(arr: Array[Int]): Boolean = {
     var res: Boolean = false
@@ -101,7 +101,7 @@ object Zed {
     res
   }
 
-  def equalWhileClass(arr: Array[Foo]): Boolean = {
+  def equalWhileClass(arr: scalaz.ImmutableArray[Foo]): Boolean = {
     var res: Boolean = false
     var i: Int = 0
     val target: Foo = Foo(10000, "hello there", true)