```
> bench/jmh:run -i 15 -wi 15 -f1 -t10 .*ClassScalaz

before
[info] EqualBench.equalDiffClassScalaz  avgt   15  3910.586 ± 55.692  ns/op
[info] EqualBench.equalSameClassScalaz  avgt   15  3913.138 ± 65.819  ns/op

after
[info] Benchmark                         Mode  Cnt        Score       Error  Units
[info] EqualBench.equalDiffClassScalaz  avgt   15  2953.313 ± 39.848  ns/op
[info] EqualBench.equalSameClassScalaz  avgt   15    43.534 ±  0.594  ns/op

c.f. cats
[info] EqualBench.equalDiffClassCats  avgt   15  7193.795 ± 117.522  ns/op
[info] EqualBench.equalSameClassCats  avgt   15    10.836 ±   0.152  ns/op
```