@fommil 2017-11-24
I'm interested in why cats appears to have a 10ns constant overhead, whereas scalaz has 40ns. Probably syntax related.

I did a memory allocation capture of the "not the same list" case and this is what yourkit says...


```
"Class","Objects","Shallow Size","Retained Size"
"scalaz.Equal$$anon$4 sun.misc.Launcher$AppClassLoader","4352679","69642864","69642864"
"scalaz.IListInstance0$$anon$4 sun.misc.Launcher$AppClassLoader","4352658","104463792","104463904"
"scalaz.syntax.EqualOps sun.misc.Launcher$AppClassLoader","4352653","104463672","104463672"
"scala.collection.immutable.$colon$colon sun.misc.Launcher$AppClassLoader","59940","1438560","302360152"
"scalaz.ICons sun.misc.Launcher$AppClassLoader","39960","959040","302360152"
"java.lang.Integer","35347","565552","565552"
"char[]","12074","953320","953320"
```

which does look like we're paying the syntax and `ListEqual` creation tax.

Incidentally, this whole problem of having to create a fresh implicit instance when there are type parameters is something I think we can solve with caching in https://gitlab.com/fommil/scalaz-deriving/issues/35