@fommil 2017-11-24
investigating the syntax tax by changing to

```
  def equalAll[A: Equal](l0: IList[A], l1: IList[A]): Boolean = Equal[IList[A]].equal(l0, l1)
```

gives

```
[info] Benchmark                        Mode  Cnt     Score    Error  Units
[info] EqualBench.equalDiffClassScalaz  avgt   15  2969.303 ± 66.155  ns/op
[info] EqualBench.equalSameClassScalaz  avgt   15    27.123 ±  0.325  ns/op
```