@fommil 2017-11-24
One more optimisation... by putting the instance based shortcut in the syntax we cover ourselves.

```
[info] EqualBench.equalDiffClassScalaz  avgt   15  2960.812 ± 38.178  ns/op
[info] EqualBench.equalSameClassScalaz  avgt   15    43.449 ±  0.550  ns/op
```