@fosskers 2017-11-24
Should this PR be left often for a bit for further experimentation?