@fommil 2017-11-24
when new scalaz is released, we can delete the monkey patch