@fosskers 2017-11-25
`master` was updated to use the new ScalaZ, and at the same time I borrow the changes here manually (modulo the auto-plugin).

From my own tests, it looks like only the `while` case with `scalaz-deriving` is acting funky. Otherwise, `scalaz-deriving` totally keeps up! The next step would be to do tests on a deeply nested ADT, like you suggested.